Reorganise pages under xdm path. Add redirect for old bookmarks.
